backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Simple todo backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["tortoise-orm>=0.25.0"]

[project.urls]
homepage = "https://github.com/ya55en/pact-showcase"

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "pytest-anyio>=0.0.0",
  "pytest-mock>=3.14.0",
  "pytest-spec>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["todo"]


[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "*_spec.py"]
python_classes = ["Test*", "Describe*"]
python_functions = ["test_*", "it_*"]
spec_test_format = "{result} {docstring_summary}"


[tool.ruff]
line-length = 96

[tool.ruff.format]
quote-style = "single"
